Update Q-learning and MCTS
Q-learning is unimplementable. I tried few hundred episodes for Qlearning player, and the .txt table is 22GB! It will crash when being loaded!
MCTS performs poorly as well.
I will dig into MCTS and approximate Q-learning tomorrow!
The Halma-main-reference is the agents research from a repo on github. They claimed to have deployed MCTS, but still need to be investigated and we shall see if we can borrow their thoughts or codes or something else.